rbac

+++ tecw_02_flask/app/blueprints/users.py
@@ -1,6 +1,8 @@
 from functools import wraps
 
-from flask import Blueprint, render_template, redirect, request, abort
+from flask import Blueprint, render_template, redirect, request, abort, session, flash
+
+from ..access_control import check_session, check_role, check_self, check_role_or_self
 from ..db import db
 from ..handle_files import save_file
 from ..models import User
@@ -36,17 +38,22 @@ def load_user(f):
 # ===================================
 
 @users_bp.route('/', methods=['GET'])
+@check_session
 def get_users():
     users = User.query.all()
     return render_template('users/users_list.html', users=users)
 
 
 @users_bp.route('/new', methods=['GET'])
+@check_session
+@check_role
 def new_user():
     return render_template('users/user_new.html')
 
 
 @users_bp.route('/', methods=['POST'])
+@check_session
+@check_role
 @save_file
 def create_user(picture=None):
     user = User(
@@ -66,18 +73,23 @@ def create_user(picture=None):
 
 
 @users_bp.route('/<int:users_id>')
+@check_session
 @load_user
 def get_user_by_id(user):
     return render_template('users/user_detail.html', user=user)
 
 
 @users_bp.route('/<int:users_id>/edit', methods=['GET'])
+@check_session
+@check_role_or_self
 @load_user
 def edit_user(user):
     return render_template('users/user_edit.html', user=user)
 
 
 @users_bp.route('/<int:users_id>/edit', methods=['POST'])
+@check_session
+@check_role_or_self
 @load_user
 @save_file
 def update_user(user, picture=None):
@@ -94,8 +106,13 @@ def update_user(user, picture=None):
 
 
 @users_bp.route('/<int:users_id>/delete', methods=['POST'])
+@check_session
+@check_role_or_self
 @load_user
 def delete_user(user):
+    if session['user']['id'] == user.id and session['user']['role'] == 'admin':
+        flash("Un administrador no puede eliminarse a sí mismo.", "error")
+        return redirect(f"/users/{user.id}")
     db.session.delete(user)
     db.session.commit()
     return redirect("/users")
